Add gridlayout as a component. Add dynamic values to convergence.tsx.
