+++ components/GridLayout/index.ts
@@ -1,0 +1,3 @@
+import GridLayout from "./GridLayout";
+
+export default GridLayout;
